feat(condo): DOMA-6834 move send ticket comment notifications to worker
