+++ apps/condo/domains/ticket/utils/handlers.js
@@ -1,7 +1,5 @@
 const dayjs = require('dayjs')
-const compact = require('lodash/compact')
 const get = require('lodash/get')
-const uniq = require('lodash/uniq')
 
 const conf = require('@open-condo/config')
 const { featureToggleManager } = require('@open-condo/featureflags/featureToggleManager')
@@ -14,15 +12,17 @@ const { md5 } = require('@condo/domains/common/utils/crypto')
 const {
     TICKET_ASSIGNEE_CONNECTED_TYPE, TICKET_EXECUTOR_CONNECTED_TYPE, TICKET_STATUS_OPENED_TYPE,
     TICKET_STATUS_IN_PROGRESS_TYPE, TICKET_STATUS_COMPLETED_TYPE, TICKET_STATUS_RETURNED_TYPE,
-    TICKET_STATUS_DECLINED_TYPE, TICKET_COMMENT_ADDED_TYPE, TRACK_TICKET_IN_DOMA_APP_TYPE,
+    TICKET_STATUS_DECLINED_TYPE, TRACK_TICKET_IN_DOMA_APP_TYPE,
 } = require('@condo/domains/notification/constants/constants')
 const { sendMessage } = require('@condo/domains/notification/utils/serverSchema')
 const { ORGANIZATION_NAME_PREFIX_AND_QUOTES_REGEXP } = require('@condo/domains/organization/constants/common')
 const { Resident } = require('@condo/domains/resident/utils/serverSchema')
-const { RESIDENT_COMMENT_TYPE } = require('@condo/domains/ticket/constants')
 const { STATUS_IDS } = require('@condo/domains/ticket/constants/statusTransitions')
+const {
+    sendTicketCommentNotifications: sendTicketCommentNotificationsTask,
+} = require('@condo/domains/ticket/tasks/sendTicketCommentNotifications')
 const { UserTicketCommentReadTime } = require('@condo/domains/ticket/utils/serverSchema')
-const { RESIDENT, STAFF } = require('@condo/domains/user/constants/common')
+const { RESIDENT } = require('@condo/domains/user/constants/common')
 
 const { Ticket, TicketCommentsTime } = require('./serverSchema')
 
@@ -262,107 +262,10 @@ const sendTicketNotifications = async (requestData) => {
         }
     }
 }
-
-const CREATE_RESIDENT_VISIBLE_COMMENT_BY_STAFF_TYPE = 'CREATE_RESIDENT_VISIBLE_COMMENT_BY_STAFF'
-const CREATE_RESIDENT_COMMENT_BY_RESIDENT_TYPE = 'CREATE_RESIDENT_COMMENT_BY_RESIDENT'
-
-const detectTicketCommentEventTypes = async ({ operation, updatedItem, ticket }) => {
-    const isCreateOperation = operation === 'create'
-    const createdById = get(updatedItem, 'createdBy')
-    const canReadByResident = get(ticket, 'canReadByResident')
-    const commentType = get(updatedItem, 'type')
-    const isResidentComment = commentType === RESIDENT_COMMENT_TYPE
-    const isCreateResidentCommentOperation = isCreateOperation && isResidentComment
-
-    const createdByUser = await getById('User', createdById)
-    const isCreatedByResident = createdByUser.type === RESIDENT
-    const isCreatedByStaff = createdByUser.type === STAFF
-
-    const result = {}
-
-    result[CREATE_RESIDENT_VISIBLE_COMMENT_BY_STAFF_TYPE] = isCreateResidentCommentOperation && canReadByResident && isCreatedByStaff
-
-    result[CREATE_RESIDENT_COMMENT_BY_RESIDENT_TYPE] = isCreateResidentCommentOperation && isCreatedByResident
-
-    return result
-}
-
 const sendTicketCommentNotifications = async (requestData) => {
-    const { updatedItem, context } = requestData
-    const ticket = await Ticket.getOne(context, { id: updatedItem.ticket })
-    const eventTypes = await detectTicketCommentEventTypes({ ...requestData, ticket })
-    const clientId = get(ticket, 'client.id')
-    const organizationId = get(ticket, 'organization.id')
-    const propertyId = get(ticket, 'property.id')
-    const unitName = get(ticket, 'unitName')
-    const unitType = get(ticket, 'unitType')
+    const { operation, updatedItem } = requestData
 
-    // no client in ticket means there is no resident connected to this ticket
-    if (!clientId) return
-
-    // TODO(DOMA-2822): get rid of this extra request by returning country within nested organization data
-    const organization = await getByCondition('Organization', {
-        id: organizationId,
-        deletedAt: null,
-    })
-    const lang = get(COUNTRIES, [organization.country, 'locale'], conf.DEFAULT_LOCALE)
-
-    const resident = await Resident.getOne(context, {
-        user: { id: clientId },
-        property: { id: propertyId },
-        organization: { id: organizationId },
-        unitName,
-        unitType,
-        deletedAt: null,
-    })
-
-    if (eventTypes[CREATE_RESIDENT_VISIBLE_COMMENT_BY_STAFF_TYPE] && clientId) {
-        await sendMessage(context, {
-            lang,
-            to: { user: { id: clientId } },
-            type: TICKET_COMMENT_ADDED_TYPE,
-            meta: {
-                dv: 1,
-                data: {
-                    ticketId: ticket.id,
-                    ticketNumber: ticket.number,
-                    userId: clientId,
-                    commentId: updatedItem.id,
-                    url: `${conf.SERVER_URL}/ticket/${ticket.id}`,
-                    residentId: get(resident, 'id', null),
-                },
-            },
-            sender: updatedItem.sender,
-            organization: { id: organization.id },
-        })
-    }
-
-    if (eventTypes[CREATE_RESIDENT_COMMENT_BY_RESIDENT_TYPE]) {
-        const ticketExecutorUserId = get(ticket, 'executor.id')
-        const ticketAssigneeUserId = get(ticket, 'assignee.id')
-        const usersToSendMessage = uniq(compact([ticketExecutorUserId, ticketAssigneeUserId]))
-
-        for (const userId of usersToSendMessage) {
-            await sendMessage(context, {
-                lang,
-                to: { user: { id: userId } },
-                type: TICKET_COMMENT_ADDED_TYPE,
-                meta: {
-                    dv: 1,
-                    data: {
-                        ticketId: ticket.id,
-                        ticketNumber: ticket.number,
-                        userId: userId,
-                        commentId: updatedItem.id,
-                        url: `${conf.SERVER_URL}/ticket/${ticket.id}`,
-                        residentId: get(resident, 'id', null),
-                    },
-                },
-                sender: updatedItem.sender,
-                organization: { id: organization.id },
-            })
-        }
-    }
+    await sendTicketCommentNotificationsTask.delay({ operation, updatedItem })
 }
 
 const createOrUpdateTicketCommentsTime = async (context, updatedItem, userType) => {
@@ -447,7 +350,6 @@ module.exports = {
     sendTicketNotifications,
     sendTicketCommentNotifications,
     detectTicketEventTypes,
-    detectTicketCommentEventTypes,
     createOrUpdateTicketCommentsTime,
     updateTicketLastCommentTime,
     ASSIGNEE_CONNECTED_EVENT_TYPE,
